backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy
PyJWT
bcrypt==4.1.1
email-validator==2.1.0
qrcode
Pillow
